+++ Examples/Basic iOS Example/iCarouselExampleAppDelegate.m
@@ -16,6 +16,7 @@
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
+    window = [[UIWindow alloc] initWithFrame:[UIScreen mainScreen].bounds];
     window.rootViewController = viewController;
     [window makeKeyAndVisible];
     return YES;